% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{add_constant}
\alias{add_constant}
\title{Add constant function}
\usage{
add_constant(X)
}
\arguments{
\item{X}{X dataset}
}
\value{
X dataset with the constant
}
\description{
Process of adding a constant into the dataframe
}
\author{
Alexandre Rives, NDiaye Deffa, Frintz Elisa
}
