% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{AIC_function}
\alias{AIC_function}
\title{AIC}
\usage{
AIC_function(LL, df)
}
\arguments{
\item{LL}{the Log-Likelihood}

\item{df}{the dataframe of the model}
}
\value{
the AIC of the model
}
\description{
The function that return the Akaike Information Criterion (AIC)
}
\author{
Alexandre Rives, NDiaye Deffa, Frintz Elisa
}
