% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_function.R
\name{batch_gradient_descent}
\alias{batch_gradient_descent}
\title{Batch gradient descent}
\usage{
batch_gradient_descent(df, var_X, var_y, learning_rate, max_iter)
}
\arguments{
\item{df}{dataset}

\item{var_X}{names of the X columns}

\item{var_y}{names of the Y columns}

\item{learning_rate}{the learning rate}

\item{max_iter}{number of iterations}
}
\value{
list of theta and the cost list
}
\description{
Batch gradient descent computes the gradient using the whole dataset
}
\author{
Frintz Elisa, NDiaye Deffa, Rives Alexandre
}
