% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breast.R
\docType{data}
\name{breast}
\alias{breast}
\title{Breast cancer}
\format{
A data frame with 699 rows and 10 features:
\describe{
\item{clump}{Clump Thickness}
\item{ucellsize}{Uniformity of Cell Size}
\item{ucellshape}{Uniformity of Cell Shape}
\item{mgadhesion}{Marginal Adhesion}
\item{sepics}{Single Epithelial Cell Size}
\item{bnuclei}{Bare Nuclei}
\item{bchromatin}{Bland Chromatin}
\item{normnucl}{Normal Nucleoli}
\item{mitoses}{Mitoses}
\item{class}{0-1}
}
}
\usage{
breast
}
\description{
Breast cancer
}
\keyword{datasets}
