% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{df_mini_batch}
\alias{df_mini_batch}
\title{Mini batch}
\usage{
df_mini_batch(df, df_initial, nb_batch)
}
\arguments{
\item{df}{dataset without the last mini batch}

\item{df_initial}{full dataset}

\item{nb_batch}{size of the batch}
}
\value{
a mini batch
}
\description{
Take a number of lines of the dataset
}
\author{
Alexandre Rives, NDiaye Deffa, Frintz Elisa
}
