% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{dummies_y}
\alias{dummies_y}
\title{Re-coding target variable}
\usage{
dummies_y(y)
}
\arguments{
\item{y}{data of the target variable}
}
\value{
A re-coding quantitative data of the target variable.
}
\description{
Process of returning a data set with re-coding the target variable
}
\author{
Frintz Elisa, NDiaye Deffa, Rives Alexandre
}
