% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{f_Formula}
\alias{f_Formula}
\title{Formula function}
\usage{
f_Formula(formula, data)
}
\arguments{
\item{formula}{the logistic regression equation}

\item{data}{A data frame or matrix}
}
\value{
a filtered dataframe
}
\description{
Function that return a filtered dataframe according to the equation given by the user.
}
\author{
Frintz Elisa, NDiaye Deffa, Rives Alexandre
}
