% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{filtered_x}
\alias{filtered_x}
\title{Filter feature variables}
\usage{
filtered_x(df, x)
}
\arguments{
\item{df}{dataframe}

\item{x}{feature variables}
}
\value{
A data set filtered with the feature variables.
}
\description{
Process of returning a data set filtered by the feature variables
}
\author{
Alexandre Rives
}
