% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_function.R
\name{fit}
\alias{fit}
\title{Fit function}
\usage{
fit(formula, data, ncores, batch_size, normalize)
}
\arguments{
\item{x}{A data frame or matrix}
}
\value{
A trained dataset
}
\description{
Function that return a trained dataset using the gradient descent.
}
\examples{
fit(y ~ ., x, ncores = 4, normalize = TRUE)

}
\author{
Alexandre Rives
}
