% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{log_loss_function}
\alias{log_loss_function}
\title{Log Loss function}
\usage{
log_loss_function(y_pred, y)
}
\arguments{
\item{y_pred}{the y predicted}

\item{y}{targets}
}
\value{
the mean of the losses
}
\description{
The function that return
}
\author{
Alexandre Rives, NDiaye Deffa, Frintz Elisa
}
