% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{logLikelihood_function}
\alias{logLikelihood_function}
\title{logLikelihood}
\usage{
logLikelihood_function(coef, df)
}
\arguments{
\item{coef}{coefficients of the model}

\item{df}{a dataframe}
}
\value{
the Log-Likelihood
}
\description{
The function that return the logLikelihood
}
\author{
Alexandre Rives, NDiaye Deffa, Frintz Elisa
}
