% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\docType{data}
\name{ncores}
\alias{ncores}
\title{Get the cores of the CPU}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ncores
}
\arguments{
\item{n_cores}{A number of cores}
}
\value{
The number of cores.
}
\description{
Process of detecting the number of cores in the CPU and return it to the function.
}
\author{
Alexandre Rives
}
\keyword{datasets}
