% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_function.R
\docType{data}
\name{norm}
\alias{norm}
\title{Fit function}
\format{
An object of class \code{logical} of length 1.
}
\usage{
norm
}
\arguments{
\item{formula}{the data frame}

\item{data}{the data frame}

\item{ncores}{the number of cores to run the fit function faster. It's better to use it with more than 10 000 data.}

\item{batch_size}{the number of batch for the mini batch mode.}

\item{normalize}{TRUE if you want to scale your feature variables. It doesn't take the qualitative data.}
}
\value{
A fitted dataset
}
\description{
Function that return a trained dataset using the gradient descent.
}
\author{
Alexandre Rives
}
\keyword{datasets}
