% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{A data frame or matrix}
}
\value{
A standardize data set
}
\description{
Process of putting different variables on the same scale
}
\author{
Alexandre Rives
}
