% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_function.R
\name{online_stochastic_gradient_descent}
\alias{online_stochastic_gradient_descent}
\title{Online stochastic gradient descent}
\usage{
online_stochastic_gradient_descent(
  df,
  var_X,
  var_y,
  learning_rate,
  max_iter,
  graph,
  epsilon
)
}
\arguments{
\item{df}{dataset}

\item{var_X}{names of the X columns}

\item{var_y}{names of the Y columns}

\item{learning_rate}{the learning rate}

\item{max_iter}{number of iterations}

\item{graph}{TRUE if you want to plot the cost list while the gradient descent is running.}

\item{epsilon}{Tolerance's threshold of the cost list convergence.}
}
\value{
list of theta and the cost list
}
\description{
Online gradient descent computes the gradient by row using a sampled dataset at each iteration
}
\author{
Frintz Elisa, NDiaye Deffa, Rives Alexandre
}
