% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_function.R
\name{predict_reg_log}
\alias{predict_reg_log}
\title{Predict function}
\usage{
predict_reg_log(object, newdata, type)
}
\arguments{
\item{object}{class object created in the fit function}

\item{newdata}{test set}

\item{type}{probs or belonging class}
}
\value{
A list
}
\description{
Function that return the predicted probs or the belonging class
}
\author{
Frintz Elisa, NDiaye Deffa, Rives Alexandre
}
