% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_function.R
\name{print.reg_log_parallel}
\alias{print.reg_log_parallel}
\title{Print function for parallel batch mode}
\usage{
\method{print}{reg_log_parallel}(x, ...)
}
\arguments{
\item{x}{the fitted object}

\item{...}{other params}
}
\description{
Function that print the elements of the fitted object
}
\author{
Frintz Elisa, NDiaye Deffa, Rives Alexandre
}
