% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{residuals_summary_function}
\alias{residuals_summary_function}
\title{Residuals summary}
\usage{
residuals_summary_function(residuals)
}
\arguments{
\item{residuals}{residuals list}
}
\value{
residuals summary
}
\description{
The function that return the residuals summary
}
\author{
Alexandre Rives, NDiaye Deffa, Frintz Elisa
}
