% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{sampled_df}
\alias{sampled_df}
\title{Sample a dataset}
\usage{
sampled_df(df)
}
\arguments{
\item{df}{dataset}
}
\value{
A sampled dataset.
}
\description{
Process of returning a sampled dataset
}
\author{
Alexandre Rives
}
