% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{dataset}
}
\value{
sigmoid
}
\description{
Formula of the sigmoid function
}
\author{
Alexandre Rives, NDiaye Deffa, Frintz Elisa
}
