% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_function.R
\name{summary.reg_log}
\alias{summary.reg_log}
\title{Fit function}
\usage{
\method{summary}{reg_log}(object, ...)
}
\arguments{
\item{object}{the fitted object}

\item{...}{other params}
}
\description{
Function that print the summary of the fitted object
}
\author{
Frintz Elisa, NDiaye Deffa, Rives Alexandre
}
