% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{dummies}
\alias{dummies}
\title{Re-coding variables}
\usage{
dummies(x)
}
\arguments{
\item{x}{A data frame or matrix}
}
\value{
A re-coding quantitative data of the dataset.
}
\description{
Process of returning a data set with re-coding feature variables
}
\author{
Alexandre Rives
}
