% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_function.R
\name{f_Formula}
\alias{f_Formula}
\title{Formula function}
\usage{
f_Formula(formula, donnee)
}
\arguments{
\item{formula}{the logistic regression equation}

\item{donnee}{A data frame or matrix}
}
\description{
Function that return a filtered dataframe according to the equation given by the user.
}
\author{
Deffa Ndiaye
}
