% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_function.R
\name{f_Formula}
\alias{f_Formula}
\title{Formula function}
\usage{
f_Formula(df)
}
\arguments{
\item{x}{A data frame or matrix}
}
\description{
Function that return a 2 vectors with the x and y variables separated.
}
\examples{
formula(iris)


}
\author{
Deffa Ndiaye
}
