% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{filtered_y}
\alias{filtered_y}
\title{Filter target variable}
\usage{
filtered_y(df, y)
}
\arguments{
\item{df}{A data frame or matrix}

\item{y}{data of the target variable}
}
\value{
A data set filtered by the target variable.
}
\description{
Process of returning a data set filtered by the target variable
}
\author{
Alexandre Rives
}
