% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_function.R
\name{fit}
\alias{fit}
\title{Fit function}
\usage{
fit(formula, data, ncores, batch_size, normalize)
}
\arguments{
\item{[formula, }{data, ncores, batch_size, normalize] Settings of the dataset, a data frame, the number of cores, the batch size for the mini-batch mode}
}
\value{
A trained dataset
}
\description{
Function that return a trained dataset using the gradient descent.
}
\author{
Alexandre Rives
}
