% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_function.R
\name{fit_reg_log}
\alias{fit_reg_log}
\title{Fit function}
\usage{
fit_reg_log(
  formula,
  data,
  mode,
  batch_size,
  normalize = FALSE,
  learning_rate,
  iter,
  ncores = 1
)
}
\arguments{
\item{formula}{Enter the target and the feature.}

\item{data}{the data frame.}

\item{mode}{choose between : batch, mini-batch or online.}

\item{batch_size}{the number of observations for the mini batch mode.}

\item{normalize}{TRUE if you want to scale your features. It doesn't take the qualitative data.}

\item{learning_rate}{used for the gradient descent.}

\item{iter}{number of iterations.}

\item{ncores}{the number of cores to run the fit function faster.}
}
\value{
object used to the predicted part

A fitted dataset
}
\description{
Function that return a trained dataset using the gradient descent.
}
\author{
Frintz Elisa, NDiaye Deffa, Rives Alexandre
}
