% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_function.R
\name{ncores}
\alias{ncores}
\title{Get the cores of the CPU}
\usage{
ncores(n_cores)
}
\arguments{
\item{n_cores}{A number of cores}
}
\value{
The number of cores.
}
\description{
Process of detecting the number of cores in the CPU and return it to the function.
}
\author{
Alexandre Rives
}
